+++ lib/palletjack/tool.rb
@@ -4,7 +4,7 @@ require 'optparse'
 require 'singleton'
 require 'rugged'
 
-class PalletJack
+module PalletJack
   # Superclass for PalletJack tool implementations
   #
   # Provides convenience methods for option parsing, file generation,
